# fix: speed up git clone by the image local git command.