… test/data/commands.json
… test/data/customPrCommands.json
… test/data/prCommands.json