This change enables git clone speed up.
If there is a git command in the build image, it uses its local git.
Otherwise it executes git clone via the sd-step (habitat) .

Currently git execution via the sd-step is very slow.
```
$ sudo docker run --rm local/sd-step time git --version
git version 2.13.5
real    0m 0.00s
user    0m 0.00s
sys     0m 0.00s

$ for i in `seq 1 5`; do sudo docker run --rm local/sd-step time sd-step exec core/git 'git --version'; done
git version 2.14.2
real    0m 55.64s
user    0m 22.27s
sys     0m 2.88s
git version 2.14.2
real    0m 51.70s
user    0m 23.42s
sys     0m 1.20s
git version 2.14.2
real    0m 55.62s
user    0m 23.84s
sys     0m 1.23s
git version 2.14.2
real    0m 52.44s
user    0m 23.28s
sys     0m 1.23s
git version 2.14.2
real    0m 51.81s
user    0m 23.39s
sys     0m 1.29s
```